+++ packages/ingest/forward/abis/convex-base-strategy.abi.ts
