+++ packages/ingest/forward/abis/crv-gauge.abi.ts
