+++ packages/ingest/forward/abis/strategy-base.abi.ts
